refactor: API类型声明优化调整
Former-commit-id: 0ae696c2e872fa90feba0c5df9a92391c02d3e0b

+++ src/utils/request.ts
@@ -7,7 +7,7 @@ import useStore from '@/store';
 const service = axios.create({
   baseURL: import.meta.env.VITE_APP_BASE_API,
   timeout: 50000,
-  headers: { 'Content-Type': 'application/json;charset=utf-8' }
+  headers: { 'Content-Type': 'application/json;charset=utf-8' },
 });
 
 // 请求拦截器
@@ -24,7 +24,7 @@ service.interceptors.request.use(
     }
     return config;
   },
-  error => {
+  (error: any) => {
     return Promise.reject(error);
   }
 );
@@ -33,36 +33,36 @@ service.interceptors.request.use(
 service.interceptors.response.use(
   (response: AxiosResponse) => {
     const { code, msg } = response.data;
-    if (code) {
-      // 有状态码判断是否为00000,除此皆为异常响应
-      if (code === '00000') {
-        return response.data;
-      } else {
-        ElMessage({
-          message: response.data.msg || '系统出错',
-          type: 'error'
-        });
-        return Promise.reject(new Error(msg || 'Error'));
+    if (code === '00000') {
+      return response.data;
+    } else {
+      // 响应数据为二进制流处理(Excel导出)
+      if (response.data instanceof ArrayBuffer) {
+        return response;
       }
-    } else {
-      // 无状态码响应直接返回
-      console.log('response', response);
-      return response;
-    }
-  },
-  error => {
-    const { code, msg } = error.response.data;
-    if (code === 'A0230') {
-      // token 过期
-      localStorage.clear(); // 清除浏览器全部缓存
-      ElMessageBox.alert('当前页面已失效，请重新登录', '提示', {});
-    } else {
+
       ElMessage({
         message: msg || '系统出错',
-        type: 'error'
+        type: 'error',
       });
       return Promise.reject(new Error(msg || 'Error'));
     }
+  },
+  (error: any) => {
+    if (error.response.data) {
+      const { code } = error.response.data;
+      // token 过期,重新登录
+      if (code === 'A0230') {
+        ElMessageBox.confirm('当前页面已失效，请重新登录', 'Warning', {
+          confirmButtonText: 'OK',
+          type: 'warning',
+        }).then(() => {
+          localStorage.clear();
+          window.location.href = '/';
+        });
+      }
+    }
+    return Promise.reject(error.message);
   }
 );
